Move translations for place subclasses to separate package
