+++ local_modules/qwant-maps-common/.gitignore
@@ -1,0 +1,1 @@
+/dist
