+++ local_modules/qwant-maps-common/i18n/br/resource.json
@@ -1,0 +1,243 @@
+{
+  "accessories shop": "stal rikoù",
+  "airport": "aerborzh",
+  "liquor store": "stal likorioù",
+  "alpine hut": "lojell alpinek",
+  "antiques shop": "stal an traoù kozh",
+  "aquarium": "akwariom",
+  "archery": "gwaregva",
+  "art shop": "stal oberennoù arzel",
+  "arts centre": "kreizenn an arzoù",
+  "athletics": "atleterezhioù",
+  "ATM": "Mekanik bilhedoù",
+  "attraction": "diduañs",
+  "bag shop": "stal seier",
+  "bakery": "bouloñjerezh",
+  "bank": "ti-bank",
+  "bar": "tavarn",
+  "basin": "poull",
+  "basketball": "basketball",
+  "bbq": "bbq",
+  "beauty salon": "saloñs braventez",
+  "bed shop": "stal gweleoù",
+  "bed and breakfast": "kambr ostizien",
+  "beverages shop": "stal evajoù",
+  "bicycle shop": "stal beloioù",
+  "biergarten": "stal bier",
+  "billiards": "bilhiardoù",
+  "bmx": "bmx",
+  "books shop": "stal levrioù",
+  "border control": "kontroll an harzoù",
+  "boules": "bouloù",
+  "boutique": "boutik",
+  "bowls": "bouloù",
+  "brewery": "braserezh artizanel",
+  "brownfield": " fraost industriel",
+  "bus station": "harz bus",
+  "bus stop": "harz bus",
+  "butcher": "boser",
+  "cafe": "kafedi",
+  "camp site": "tachenn gampiñ",
+  "canoe": "kanoe",
+  "car shop": "stal kirri",
+  "car parts shop": "stal pezhioù kirri",
+  "car repair": "stal kempenn kirri",
+  "car rental": "stal feurmiñ kirri",
+  "car wash station": "porzh gwalc'hiñ",
+  "caravan site": "tachenn garavan",
+  "carpenter": "kalvez",
+  "carpet shop": "stal tapisoù",
+  "castle": "kastell",
+  "catering": "tineller",
+  "cemetery": "bered",
+  "holiday cottage": " ti-vakañsoù",
+  "charging station": "porzh adkargañ",
+  "charity store": "stal an aluzon",
+  "chemist": "lezapotikerezh",
+  "cheese shop": "fourmajer",
+  "chess": "c'hoari echedoù",
+  "childcare centre": "magouri",
+  "chocolate shop": "stal chokolad",
+  "cinema": "sinema",
+  "climbing": "krimpadenn",
+  "climbing adventure": "avantur krimpadenn",
+  "clinic": "klinik",
+  "clothes shop": "stal dilhad",
+  "coffee shop": "stal kafe",
+  "college": "skolaj",
+  "community centre": "kreizenn gumuniezh",
+  "computer shop": "stal urzhiataerezh",
+  "confectionery shop": "stal koñfizerezh",
+  "convenience store": "ispisiri",
+  "copyshop": "stal liestennañ",
+  "cosmetics shop": "stal kosmetekerezh",
+  "courthouse": "lez-varn",
+  "coworking space": "spas-labourat boutin",
+  "cricket": "kriket",
+  "cycling": "kezeg-houarn",
+  "deli": "deli",
+  "dentist": "dentour",
+  "department store": "stal bras",
+  "dock": "kae",
+  "doctors office": "stal an doktor",
+  "dog park": "park chas",
+  "dog racing": "redadegoù chas",
+  "doityourself store": "stal bitellat",
+  "driving school": "skol-vleniañ",
+  "dry cleaning": "naetaerezh",
+  "electronics shop": "stal traoù elektronik",
+  "embassy": "kannati",
+  "equestrian": "kreizenn varc'hegezh",
+  "erotic shop": "stal traoù erotek",
+  "escape game": "c'hoari achapiñ",
+  "estate agent": "ajañs-tiez",
+  "fabric": "labouradeg",
+  "farm shop": "produioù ar feurm",
+  "fast food": "pretierezh prim",
+  "fire station": "ti ar bomperien",
+  "fitness centre": "ti-jiminas yac'haat",
+  "florist": "stal bleunioù",
+  "food court": "takad debriñ",
+  "free flying": "free flying",
+  "frozen food": "boued gourskornet",
+  "fuel station": "stal dour-tan",
+  "funeral home": "liderezh-kañv",
+  "furniture shop": "stal arrebeuri",
+  "gallery": "diskouezva",
+  "garden": "liorzh",
+  "garden centre": "kreizenn liorzhaouiñ",
+  "general store": "magazenn hollek",
+  "gift shop": "stal profoù",
+  "golf": "golf",
+  "golf course": "tachenn golf",
+  "grave yard": "bered",
+  "greengrocer": "stal legumaj",
+  "guest house": "ti ostizien",
+  "gymnastics": "ti-jiminas",
+  "hackerspace": "hackerspace",
+  "hairdresser": "ficher-blev",
+  "halt": "harz",
+  "hardware shop": "kinkailherezh",
+  "hearing aids shop": "stal benvegoù selaou",
+  "hifi shop": "stal hifi",
+  "horse racing": "redadegoù kezeg",
+  "hospital": "ospital",
+  "hostel": "ti ar re yaouank",
+  "hotel": "leti",
+  "houseware shop": "reizhoù al lastez-ti",
+  "ice cream": "stal diennoù-skorn",
+  "ice rink": "poull-rinklañ",
+  "interior decoration shop": "stal kinkladurioù diabarzh",
+  "jewelry shop": "bravigerezh",
+  "karting": "karrigoù-tan",
+  "kindergarten or preschool": "liorzh bugale pe rak-skol",
+  "kiosk": "kiosk",
+  "kitchen manufacturer": "gwerzher keginoù",
+  "lamps shop": "stal lampoù",
+  "laundry": " ti-kannañ dilhad",
+  "library": "levraoueg",
+  "long jump": "lamm hir",
+  "mall": "kreizenn genwerzh",
+  "marina": "marina",
+  "marketplace": "marc'had",
+  "massage shop": "stal dorloerezh",
+  "miniature golf": "golfig",
+  "mobile phone shop": "stal hezougelloù",
+  "model aerodrome": "nijva evit gobarioù",
+  "monument": "monumant",
+  "motel": "motel",
+  "motocross": "marc'hoù-tan kros",
+  "motor": "keflusker",
+  "motorcycle shop": "stal marc'hoù-tan",
+  "multi": "multi",
+  "museum": "mirdi",
+  "music shop": "stal sonerezh",
+  "music school": "skol sonerezh",
+  "musical instrument shop": "stal binviji sonerezh",
+  "newsagent shop": "stal kazetennerezh",
+  "nightclub": "boest-noz",
+  "nursing home": "ti ar re gozh",
+  "optician": "luneder",
+  "orienteering": "redadegoù-reteriñ",
+  "outdoor": "stal sportoù war ar maez",
+  "paragliding": "bolc'haskell",
+  "park": "park",
+  "car park": "park kiri",
+  "pastry shop": "stal pastezerezh",
+  "perfumery": "porfumerezh",
+  "pet store": "stal-loened",
+  "pharmacy": "apotikerezh",
+  "photo shop": "stal fotoioù",
+  "picnic site": "tachenn biknik",
+  "place of worship": "lec'h azeulerezh",
+  "plumber": "plomer",
+  "police": "komiserdi",
+  "polling station": "burev votiñ",
+  "post office": "ti-post",
+  "prison": "toull-bac'h",
+  "pub": "bruderezh",
+  "public building": "savadur foran",
+  "rc car": "karr pellurzhiet",
+  "recycling": "adaozañ",
+  "reservoir": "beol",
+  "restaurant": "predti",
+  "rowing": "roeñverezh",
+  "ruins": "dismantroù",
+  "running": "redadeg war-droad",
+  "sailing": "bageal",
+  "school": "skol",
+  "scuba diving": "splujerezh danvor",
+  "fishmonger": "stal gwerzhañ pesked",
+  "resale shop": "stal adwerzhañ",
+  "shelter": "gwasked",
+  "shoemaker": "kere",
+  "shoes": "botoù",
+  "shooting": "tennañ",
+  "skateboard": "skateboard",
+  "skating": "skating",
+  "skiing": "skiañ",
+  "soccer": "mell-droad",
+  "sports shop": "stal sport",
+  "sports centre": "kreizenn sport",
+  "stadium": "stad",
+  "station": "savlec'h",
+  "stationery shop": "stal paperaerezh",
+  "subway": "tren-buzhug",
+  "supermarket": "gourmarc'had",
+  "swimming pool": "poull-neuial",
+  "swimming area": "zonenn neuial",
+  "swimming pool equipment": "stal gwerzhañ poulloù-neuial",
+  "table tennis": "taol dennis",
+  "tailor": "tailher",
+  "tattoo salon": "saloñs tatou",
+  "tennis": "tennis",
+  "theatre": "c'hoariva",
+  "theme park": "park gant un tem",
+  "ticket shop": "bilhederezh",
+  "tobacco shop": "stal butun",
+  "toll booth": "treizhaj",
+  "townhall": "ti-kêr",
+  "toys shop": "stal c'hoarioù",
+  "tram stop": "arsav tram",
+  "travel agency": "ajañs-veaj",
+  "university": "skol-veur",
+  "variety store": "stal ur priz",
+  "vehicle inspection": "kontroll teknikel",
+  "veterinary": "mezeg-loened",
+  "videos shop": "stal videoioù",
+  "video games shop": "stal c'hoarioù-video",
+  "viewpoint": "gweledva",
+  "watches shop": "stal eurieroù",
+  "water park": "park dour",
+  "weapons shop": "stal armoù",
+  "warehouse club": "braswerzher",
+  "wine shop": "stal gwin",
+  "winery": "etrepaouez gwin",
+  "winter sports resort": " kreizenn sportoù ar goañv",
+  "yoga": "yoga",
+  "zoo": "zoo",
+  "city": "kêr",
+  "country": "bro",
+  "address": "chomlec'h",
+  "street": "straed"
+}
