+++ local_modules/qwant-maps-common/i18n/ca/resource.json
@@ -1,0 +1,243 @@
+{
+  "accessories shop": "botiga d'accessoris",
+  "airport": "aeroport",
+  "liquor store": "licoreria",
+  "alpine hut": "alberg de muntanya",
+  "antiques shop": "botiga d'antiguitats",
+  "aquarium": "aquari",
+  "archery": "tir amb arc",
+  "art shop": "botiga d'art",
+  "arts centre": "centre d'arts",
+  "athletics": "atletisme",
+  "ATM": "caixer automàtic",
+  "attraction": "atracció",
+  "bag shop": "botiga de bosses",
+  "bakery": "forn de pa",
+  "bank": "banc",
+  "bar": "bar de copes",
+  "basin": "palangana",
+  "basketball": "basquetbol",
+  "bbq": "barbacoa",
+  "beauty salon": "saló de bellesa",
+  "bed shop": "botiga de matalassos",
+  "bed and breakfast": "llit i esmorzar",
+  "beverages shop": "botiga de begudes",
+  "bicycle shop": "botiga de bicicletes",
+  "biergarten": "cerveseria a l'aire lliure",
+  "billiards": "billars",
+  "bmx": "bmx",
+  "books shop": "llibreria",
+  "border control": "duana",
+  "boules": "esports de boles",
+  "boutique": "boutique",
+  "bowls": "bitlles",
+  "brewery": "cerveseria",
+  "brownfield": "zona industrial abandonada",
+  "bus station": "estació de busos",
+  "bus stop": "parada d'autobús",
+  "butcher": "carnisseria",
+  "cafe": "cafè",
+  "camp site": "lloc d'acampada",
+  "canoe": "canoa",
+  "car shop": "botiga de cotxes",
+  "car parts shop": "botiga de peces de cotxes",
+  "car repair": "reparació de cotxes",
+  "car rental": "lloguer de cotxes",
+  "car wash station": "estació de rentat de cotxes",
+  "caravan site": "lloc de caravanes",
+  "carpenter": "fuster",
+  "carpet shop": "botiga de catifes",
+  "castle": "castell",
+  "catering": "restauració",
+  "cemetery": "cementiri",
+  "holiday cottage": "casa rural",
+  "charging station": "estació de càrrega",
+  "charity store": "botiga de caritat",
+  "chemist": "adrogueria",
+  "cheese shop": "formatgeria",
+  "chess": "escacs",
+  "childcare centre": "llar d'infants",
+  "chocolate shop": "botiga de xocolata",
+  "cinema": "cinema",
+  "climbing": "escalada",
+  "climbing adventure": "aventura d'escalada",
+  "clinic": "clínica",
+  "clothes shop": "botiga de roba",
+  "coffee shop": "botiga de cafè",
+  "college": "institut",
+  "community centre": "centre comunitari",
+  "computer shop": "botiga d'informàtica",
+  "confectionery shop": "merceria",
+  "convenience store": "botiga de conveniència",
+  "copyshop": "copisteria",
+  "cosmetics shop": "botiga de cosmètics",
+  "courthouse": "jutjat",
+  "coworking space": "espai de cotreball",
+  "cricket": "criquet",
+  "cycling": "bicicletes",
+  "deli": "xarcuteria",
+  "dentist": "dentista",
+  "department store": "grans magatzems",
+  "dock": "moll",
+  "doctors office": "l'oficina del doctor",
+  "dog park": "parc caní",
+  "dog racing": "carreres de gossos",
+  "doityourself store": "botiga de bricolatge",
+  "driving school": "autoescola",
+  "dry cleaning": "neteja en sec",
+  "electronics shop": "botiga d'electrònica",
+  "embassy": "ambaixada",
+  "equestrian": "centre hípic",
+  "erotic shop": "botiga eròtica",
+  "escape game": "joc d'escapada",
+  "estate agent": "agent immobiliari",
+  "fabric": "fàbrica",
+  "farm shop": "botiga de granja",
+  "fast food": "menjar ràpid",
+  "fire station": "parc de bombers",
+  "fitness centre": "gimnàs",
+  "florist": "floristeria",
+  "food court": "zona de menjars",
+  "free flying": "vol lliure",
+  "frozen food": "aliments congelats",
+  "fuel station": "benzinera",
+  "funeral home": "funerària",
+  "furniture shop": "botiga de mobles",
+  "gallery": "galeria d'art",
+  "garden": "jardí",
+  "garden centre": "centre de jardineria",
+  "general store": "magatzem general",
+  "gift shop": "botiga de regals",
+  "golf": "golf",
+  "golf course": "curs de golf",
+  "grave yard": "cementiri",
+  "greengrocer": "fruiteria",
+  "guest house": "posada",
+  "gymnastics": "gimnàs",
+  "hackerspace": "hackerspace",
+  "hairdresser": "perruqueria",
+  "halt": "aturar",
+  "hardware shop": "ferreteria",
+  "hearing aids shop": "botiga d'audiòfons",
+  "hifi shop": "botiga de productes d'àudio d'alta fidelitat",
+  "horse racing": "curses de cavalls",
+  "hospital": "hospital",
+  "hostel": "alberg",
+  "hotel": "hotel",
+  "houseware shop": "botiga d'articles per a la llar",
+  "ice cream": "gelat",
+  "ice rink": "pista de gel",
+  "interior decoration shop": "botiga de decoració d'interiors",
+  "jewelry shop": "joieria",
+  "karting": "kàrting",
+  "kindergarten or preschool": "llar d'infants",
+  "kiosk": "quiosc",
+  "kitchen manufacturer": "fabricant de cuines",
+  "lamps shop": "botiga de llums",
+  "laundry": "bugaderia",
+  "library": "biblioteca",
+  "long jump": "salt de llargada",
+  "mall": "centre comercial",
+  "marina": "port esportiu",
+  "marketplace": "mercat",
+  "massage shop": "botiga de massatges",
+  "miniature golf": "mingolf",
+  "mobile phone shop": "botiga de mòbils",
+  "model aerodrome": "aeromodelisme",
+  "monument": "monument",
+  "motel": "motel",
+  "motocross": "motocròs",
+  "motor": "motor",
+  "motorcycle shop": "botiga de motocicletes",
+  "multi": "multi",
+  "museum": "museu",
+  "music shop": "botiga de música",
+  "music school": "escola de música",
+  "musical instrument shop": "botiga d'instruments musicals",
+  "newsagent shop": "quiosc de premsa",
+  "nightclub": "club nocturn",
+  "nursing home": "residència de gent gran",
+  "optician": "òptic",
+  "orienteering": "orientació",
+  "outdoor": "a l'aire lliure",
+  "paragliding": "parapent",
+  "park": "parc",
+  "car park": "aparcament",
+  "pastry shop": "pastisseria",
+  "perfumery": "perfumeria",
+  "pet store": "botiga d'animals",
+  "pharmacy": "farmàcia",
+  "photo shop": "botiga de fotografia",
+  "picnic site": "lloc de pícnic",
+  "place of worship": "lloc de pregària",
+  "plumber": "fontaneria",
+  "police": "policia",
+  "polling station": " col·legi electoral",
+  "post office": "oficina postal",
+  "prison": "presó",
+  "pub": "pub o bar",
+  "public building": "edifici públic",
+  "rc car": "cotxe rc",
+  "recycling": "reciclatge",
+  "reservoir": "embassament",
+  "restaurant": "restaurant",
+  "rowing": "remar",
+  "ruins": "runes",
+  "running": "corrent",
+  "sailing": "navegació",
+  "school": "escola",
+  "scuba diving": "submarinisme",
+  "fishmonger": "peixateria",
+  "resale shop": "botiga de revenda",
+  "shelter": "refugi",
+  "shoemaker": "sabateria",
+  "shoes": "sabates",
+  "shooting": "tiroteig",
+  "skateboard": "skateboard",
+  "skating": "skating",
+  "skiing": "esquiar",
+  "soccer": "futbol",
+  "sports shop": "botiga d'esports",
+  "sports centre": "centre esportiu",
+  "stadium": "estadi",
+  "station": "estació",
+  "stationery shop": "papereria",
+  "subway": "metro",
+  "supermarket": "supermercat",
+  "swimming pool": "piscina",
+  "swimming area": "àrea de natació",
+  "swimming pool equipment": "piscina",
+  "table tennis": "tennis taula",
+  "tailor": "sastre",
+  "tattoo salon": "saló de tatuatge",
+  "tennis": "tennis",
+  "theatre": "teatre",
+  "theme park": "parc d'atraccions",
+  "ticket shop": "botiga d'entrades",
+  "tobacco shop": "estanc",
+  "toll booth": "cabina de peatge",
+  "townhall": "ajuntament",
+  "toys shop": "botiga de joguines",
+  "tram stop": "parada de tramvia",
+  "travel agency": "agència de viatges",
+  "university": "univeristat",
+  "variety store": "basar",
+  "vehicle inspection": "inspecció de vehicles - ITV",
+  "veterinary": "veterinari",
+  "videos shop": "videoclub",
+  "video games shop": "botiga de videojocs",
+  "viewpoint": "punt de vista",
+  "watches shop": "botiga d'ulleres",
+  "water park": "parc aquàtic",
+  "weapons shop": "armeria",
+  "warehouse club": "distribuïdor a l'engròs",
+  "wine shop": "celler",
+  "winery": "celler",
+  "winter sports resort": "estació d'esports d'hivern",
+  "yoga": "ioga",
+  "zoo": "zoo",
+  "city": "ciutat",
+  "country": "país",
+  "address": "adreça",
+  "street": "carrer"
+}
