+++ local_modules/qwant-maps-common/i18n/de/resource.json
@@ -1,0 +1,243 @@
+{
+  "accessories shop": "Zubehörshop ",
+  "airport": "Flughafen",
+  "liquor store": "Spirituosengeschäft",
+  "alpine hut": "Berghütte",
+  "antiques shop": "Antiquitätengeschäft ",
+  "aquarium": "Aquarium",
+  "archery": "Bogenschießen",
+  "art shop": "Kunstgeschäft",
+  "arts centre": "Kunstzentrum",
+  "athletics": "Leichtathletik ",
+  "ATM": "Bankautomat",
+  "attraction": "Attraktion",
+  "bag shop": "Taschengeschäft",
+  "bakery": "Bäckerei",
+  "bank": "Bank",
+  "bar": "Bar",
+  "basin": "Becken",
+  "basketball": "Basketball",
+  "bbq": "BBQ",
+  "beauty salon": "Schönheitssalon",
+  "bed shop": "Bettengeschäft",
+  "bed and breakfast": "Übernachtung mit Frühstück",
+  "beverages shop": "Getränkeladen",
+  "bicycle shop": "Fahrradladen",
+  "biergarten": "Biergarten",
+  "billiards": "Billard",
+  "bmx": "BMX",
+  "books shop": "Buchladen",
+  "border control": "Grenzkontrolle",
+  "boules": "Boule",
+  "boutique": "Boutique",
+  "bowls": "Bowlen",
+  "brewery": "Brauerei",
+  "brownfield": "Brachfläche",
+  "bus station": "Busbahnhof",
+  "bus stop": "Bushaltestelle",
+  "butcher": "Metzger",
+  "cafe": "Café",
+  "camp site": "Campingplatz",
+  "canoe": "Kanu",
+  "car shop": "KFZ-Werkstatt",
+  "car parts shop": "Geschäft für Autozubehör",
+  "car repair": "Autoreparatur",
+  "car rental": "Autovermietung",
+  "car wash station": "Waschstraße",
+  "caravan site": "Campingplatz",
+  "carpenter": "Tischler",
+  "carpet shop": "Teppichhandel",
+  "castle": "Burg",
+  "catering": "Catering",
+  "cemetery": "Friedhof",
+  "holiday cottage": "Ferienhaus",
+  "charging station": "Ladestation",
+  "charity store": "Wohltätigkeitsladen",
+  "chemist": "Drogerie",
+  "cheese shop": "Käseladen",
+  "chess": "Schach",
+  "childcare centre": "Kindertagesstätte",
+  "chocolate shop": "Schokoladengeschäft",
+  "cinema": "Kino",
+  "climbing": "Klettern",
+  "climbing adventure": "Kletterabenteuer",
+  "clinic": "Klinik",
+  "clothes shop": "Kleidergeschäft",
+  "coffee shop": "Café",
+  "college": "College",
+  "community centre": "Gemeindezentrum",
+  "computer shop": "Computergeschäft",
+  "confectionery shop": "Süßwarenladen",
+  "convenience store": "Kleiner Lebensmittelladen",
+  "copyshop": "Copyshop",
+  "cosmetics shop": "Kosmetikgeschäft",
+  "courthouse": "Gerichtsgebäude",
+  "coworking space": "Coworking-Space",
+  "cricket": "Kricket",
+  "cycling": "Rad fahren ",
+  "deli": "Feinkostladen",
+  "dentist": "Zahnarzt",
+  "department store": "Kaufhaus",
+  "dock": "Kai",
+  "doctors office": "Arztpraxis",
+  "dog park": "Hundepark",
+  "dog racing": "Hunderennen",
+  "doityourself store": "Heimwerkermarkt",
+  "driving school": "Fahrschule",
+  "dry cleaning": "Reinigung",
+  "electronics shop": "Elektronikgeschäft",
+  "embassy": "Botschaft",
+  "equestrian": "Reitsport",
+  "erotic shop": "Erotikshop",
+  "escape game": "Escape Game",
+  "estate agent": "Immobilienmakler",
+  "fabric": "Bau",
+  "farm shop": "Hofladen",
+  "fast food": "Fast Food",
+  "fire station": "Feuerwache",
+  "fitness centre": "Fitnessstudio",
+  "florist": "Blumenhändler",
+  "food court": "Gastronomiebereich",
+  "free flying": "Gratis Flug",
+  "frozen food": "Tiefkühlkost",
+  "fuel station": "Tankstelle",
+  "funeral home": "Bestattungsinstitut",
+  "furniture shop": "Möbelgeschäft",
+  "gallery": "Galerie",
+  "garden": "Garten",
+  "garden centre": "Gartenmarkt",
+  "general store": "Gemischtwarenladen",
+  "gift shop": "Geschenkartikelladen",
+  "golf": "Golf",
+  "golf course": "Golfplatz",
+  "grave yard": "Friedhof",
+  "greengrocer": "Gemüsehändler",
+  "guest house": "Pension",
+  "gymnastics": "Turnen",
+  "hackerspace": "Hackerspace",
+  "hairdresser": "Friseur",
+  "halt": "Halt",
+  "hardware shop": "Eisenwarenhandlung",
+  "hearing aids shop": "Geschäft für Hörgeräte",
+  "hifi shop": "HiFi-Fachgeschäft",
+  "horse racing": "Pferderennen",
+  "hospital": "Krankenhaus",
+  "hostel": "Jugendherberge",
+  "hotel": "Hotel",
+  "houseware shop": "Haushaltswarenladen",
+  "ice cream": "Eiscreme",
+  "ice rink": "Eisbahn",
+  "interior decoration shop": "Geschäft für Innenausstattung",
+  "jewelry shop": "Juweliergeschäft",
+  "karting": "Kartsport",
+  "kindergarten or preschool": "Kindergarten oder Vorschule",
+  "kiosk": "Kiosk",
+  "kitchen manufacturer": "Küchenhersteller",
+  "lamps shop": "Lampengeschäft",
+  "laundry": "Wäscherei",
+  "library": "Bücherei",
+  "long jump": "Weitsprung",
+  "mall": "Einkaufszentrum",
+  "marina": "Yachthafen",
+  "marketplace": "Marktplatz",
+  "massage shop": "Massagesalon",
+  "miniature golf": "Minigolf",
+  "mobile phone shop": "Handyshop",
+  "model aerodrome": "Modellflugplatz",
+  "monument": "Denkmal",
+  "motel": "Motel",
+  "motocross": "Motocross",
+  "motor": "Motor",
+  "motorcycle shop": "Motorradhändler",
+  "multi": "multi",
+  "museum": "Museum",
+  "music shop": "Musikgeschäft",
+  "music school": "Musikschule",
+  "musical instrument shop": "Musikalienhandlung",
+  "newsagent shop": "Zeitungshändler",
+  "nightclub": "Nachtclub",
+  "nursing home": "Pflegeheim",
+  "optician": "Optiker",
+  "orienteering": "Orientierungslauf",
+  "outdoor": "im Freien",
+  "paragliding": "Paragleiten",
+  "park": "Park",
+  "car park": "Parkplatz",
+  "pastry shop": "Konditorei",
+  "perfumery": "Parfümerie",
+  "pet store": " Tierhandlung",
+  "pharmacy": "Apotheke",
+  "photo shop": "Fotoladen",
+  "picnic site": "Picknickplatz",
+  "place of worship": "Gotteshaus",
+  "plumber": "Klempner",
+  "police": "Polizei",
+  "polling station": "Wahllokal",
+  "post office": "Postamt",
+  "prison": "Gefängnis",
+  "pub": "Pub",
+  "public building": "Öffentliches Gebäude",
+  "rc car": "RC-Car",
+  "recycling": "Recycling",
+  "reservoir": "Reservoir",
+  "restaurant": "Restaurant",
+  "rowing": "Rudern",
+  "ruins": "Ruinen",
+  "running": "Rennen",
+  "sailing": "Segeln",
+  "school": "Schule",
+  "scuba diving": "Sporttauchen",
+  "fishmonger": "Fischhändler",
+  "resale shop": "Wiederverkäufer",
+  "shelter": "Unterkunft",
+  "shoemaker": "Schuster",
+  "shoes": "Schuhe",
+  "shooting": "Schießen",
+  "skateboard": "Skateboard",
+  "skating": "Skaten",
+  "skiing": "Skifahren",
+  "soccer": "Fußball",
+  "sports shop": "Sportgeschäft",
+  "sports centre": "Sportzentrum",
+  "stadium": "Stadion",
+  "station": "Bahnhof",
+  "stationery shop": "Schreibwarengeschäft",
+  "subway": "Untergrundbahn",
+  "supermarket": "Supermarkt",
+  "swimming pool": "Schwimmbad",
+  "swimming area": "Badebereich",
+  "swimming pool equipment": "Swimming Pool-Bedarf",
+  "table tennis": "Tischtennis",
+  "tailor": "Schneider",
+  "tattoo salon": "Tätowierstudio ",
+  "tennis": "Tennis",
+  "theatre": "Theater",
+  "theme park": "Freizeitpark",
+  "ticket shop": "Ticketshop",
+  "tobacco shop": "Tabakgeschäft",
+  "toll booth": "Mautstation",
+  "townhall": "Rathaus",
+  "toys shop": "Spielzeuggeschäft",
+  "tram stop": "Straßenbahnhaltestelle",
+  "travel agency": "Reiseagentur",
+  "university": "Universität",
+  "variety store": "Kramladen",
+  "vehicle inspection": "Fahrzeuginspektion",
+  "veterinary": "Tierarzt",
+  "videos shop": "Videothek",
+  "video games shop": "Geschäft für Computerspiele",
+  "viewpoint": "Aussichtspunkt",
+  "watches shop": "Uhrengeschäft",
+  "water park": "Erlebnisbad",
+  "weapons shop": "Waffengeschäft",
+  "warehouse club": "Lagerhausclub",
+  "wine shop": "Weingeschäft",
+  "winery": "Weinladen",
+  "winter sports resort": "Wintersportort",
+  "yoga": "Yoga",
+  "zoo": "Zoo",
+  "city": "Stadt",
+  "country": "Land",
+  "address": "Adresse",
+  "street": "Straße"
+}
