+++ local_modules/qwant-maps-common/i18n/en/resource.json
@@ -1,0 +1,243 @@
+{
+  "accessories shop": "accessories shop",
+  "airport": "airport",
+  "liquor store": "liquor store",
+  "alpine hut": "alpine hut",
+  "antiques shop": "antiques shop",
+  "aquarium": "aquarium",
+  "archery": "archery",
+  "art shop": "art shop",
+  "arts centre": "arts centre",
+  "athletics": "athletics",
+  "ATM": "ATM",
+  "attraction": "attraction",
+  "bag shop": "bag shop",
+  "bakery": "bakery",
+  "bank": "bank",
+  "bar": "bar",
+  "basin": "basin",
+  "basketball": "basketball",
+  "bbq": "bbq",
+  "beauty salon": "beauty salon",
+  "bed shop": "bed shop",
+  "bed and breakfast": "bed and breakfast",
+  "beverages shop": "beverages shop",
+  "bicycle shop": "bicycle shop",
+  "biergarten": "biergarten",
+  "billiards": "billiards",
+  "bmx": "bmx",
+  "books shop": "books shop",
+  "border control": "border control",
+  "boules": "boules",
+  "boutique": "boutique",
+  "bowls": "bowls",
+  "brewery": "brewery",
+  "brownfield": "brownfield",
+  "bus station": "bus station",
+  "bus stop": "bus stop",
+  "butcher": "butcher",
+  "cafe": "cafe",
+  "camp site": "camp site",
+  "canoe": "canoe",
+  "car shop": "car shop",
+  "car parts shop": "car parts shop",
+  "car repair": "car repair",
+  "car rental": "car rental",
+  "car wash station": "car wash station",
+  "caravan site": "caravan site",
+  "carpenter": "carpenter",
+  "carpet shop": "carpet shop",
+  "castle": "castle",
+  "catering": "catering",
+  "cemetery": "cemetery",
+  "holiday cottage": "holiday cottage",
+  "charging station": "charging station",
+  "charity store": "charity store",
+  "chemist": "chemist",
+  "cheese shop": "cheese shop",
+  "chess": "chess",
+  "childcare centre": "childcare centre",
+  "chocolate shop": "chocolate shop",
+  "cinema": "cinema",
+  "climbing": "climbing",
+  "climbing adventure": "climbing adventure",
+  "clinic": "clinic",
+  "clothes shop": "clothes shop",
+  "coffee shop": "coffee shop",
+  "college": "college",
+  "community centre": "community centre",
+  "computer shop": "computer shop",
+  "confectionery shop": "confectionery shop",
+  "convenience store": "convenience store",
+  "copyshop": "copyshop",
+  "cosmetics shop": "cosmetics shop",
+  "courthouse": "courthouse",
+  "coworking space": "coworking space",
+  "cricket": "cricket",
+  "cycling": "cycling",
+  "deli": "deli",
+  "dentist": "dentist",
+  "department store": "department store",
+  "dock": "dock",
+  "doctors office": "doctors office",
+  "dog park": "dog park",
+  "dog racing": "dog racing",
+  "doityourself store": "doityourself store",
+  "driving school": "driving school",
+  "dry cleaning": "dry cleaning",
+  "electronics shop": "electronics shop",
+  "embassy": "embassy",
+  "equestrian": "equestrian",
+  "erotic shop": "erotic shop",
+  "escape game": "escape game",
+  "estate agent": "estate agent",
+  "fabric": "fabric",
+  "farm shop": "farm shop",
+  "fast food": "fast food",
+  "fire station": "fire station",
+  "fitness centre": "fitness centre",
+  "florist": "florist",
+  "food court": "food court",
+  "free flying": "free flying",
+  "frozen food": "frozen food",
+  "fuel station": "fuel station",
+  "funeral home": "funeral home",
+  "furniture shop": "furniture shop",
+  "gallery": "gallery",
+  "garden": "garden",
+  "garden centre": "garden centre",
+  "general store": "general store",
+  "gift shop": "gift shop",
+  "golf": "golf",
+  "golf course": "golf course",
+  "grave yard": "grave yard",
+  "greengrocer": "greengrocer",
+  "guest house": "guest house",
+  "gymnastics": "gymnastics",
+  "hackerspace": "hackerspace",
+  "hairdresser": "hairdresser",
+  "halt": "halt",
+  "hardware shop": "hardware shop",
+  "hearing aids shop": "hearing aids shop",
+  "hifi shop": "hifi shop",
+  "horse racing": "horse racing",
+  "hospital": "hospital",
+  "hostel": "hostel",
+  "hotel": "hotel",
+  "houseware shop": "houseware shop",
+  "ice cream": "ice cream",
+  "ice rink": "ice rink",
+  "interior decoration shop": "interior decoration shop",
+  "jewelry shop": "jewelry shop",
+  "karting": "karting",
+  "kindergarten or preschool": "kindergarten or preschool",
+  "kiosk": "kiosk",
+  "kitchen manufacturer": "kitchen manufacturer",
+  "lamps shop": "lamps shop",
+  "laundry": "laundry",
+  "library": "library",
+  "long jump": "long jump",
+  "mall": "mall",
+  "marina": "marina",
+  "marketplace": "marketplace",
+  "massage shop": "massage shop",
+  "miniature golf": "miniature golf",
+  "mobile phone shop": "mobile phone shop",
+  "model aerodrome": "model aerodrome",
+  "monument": "monument",
+  "motel": "motel",
+  "motocross": "motocross",
+  "motor": "motor",
+  "motorcycle shop": "motorcycle shop",
+  "multi": "multi",
+  "museum": "museum",
+  "music shop": "music shop",
+  "music school": "music school",
+  "musical instrument shop": "musical instrument shop",
+  "newsagent shop": "newsagent shop",
+  "nightclub": "nightclub",
+  "nursing home": "nursing home",
+  "optician": "optician",
+  "orienteering": "orienteering",
+  "outdoor": "outdoor",
+  "paragliding": "paragliding",
+  "park": "park",
+  "car park": "car park",
+  "pastry shop": "pastry shop",
+  "perfumery": "perfumery",
+  "pet store": "pet store",
+  "pharmacy": "pharmacy",
+  "photo shop": "photo shop",
+  "picnic site": "picnic site",
+  "place of worship": "place of worship",
+  "plumber": "plumber",
+  "police": "police",
+  "polling station": "polling station",
+  "post office": "post office",
+  "prison": "prison",
+  "pub": "pub",
+  "public building": "public building",
+  "rc car": "rc car",
+  "recycling": "recycling",
+  "reservoir": "reservoir",
+  "restaurant": "restaurant",
+  "rowing": "rowing",
+  "ruins": "ruins",
+  "running": "running",
+  "sailing": "sailing",
+  "school": "school",
+  "scuba diving": "scuba diving",
+  "fishmonger": "fishmonger",
+  "resale shop": "resale shop",
+  "shelter": "shelter",
+  "shoemaker": "shoemaker",
+  "shoes": "shoes",
+  "shooting": "shooting",
+  "skateboard": "skateboard",
+  "skating": "skating",
+  "skiing": "skiing",
+  "soccer": "soccer",
+  "sports shop": "sports shop",
+  "sports centre": "sports centre",
+  "stadium": "stadium",
+  "station": "station",
+  "stationery shop": "stationery shop",
+  "subway": "subway",
+  "supermarket": "supermarket",
+  "swimming pool": "swimming pool",
+  "swimming area": "swimming area",
+  "swimming pool equipment": "swimming pool equipment",
+  "table tennis": "table tennis",
+  "tailor": "tailor",
+  "tattoo salon": "tattoo salon",
+  "tennis": "tennis",
+  "theatre": "theatre",
+  "theme park": "theme park",
+  "ticket shop": "ticket shop",
+  "tobacco shop": "tobacco shop",
+  "toll booth": "toll booth",
+  "townhall": "townhall",
+  "toys shop": "toys shop",
+  "tram stop": "tram stop",
+  "travel agency": "travel agency",
+  "university": "university",
+  "variety store": "variety store",
+  "vehicle inspection": "vehicle inspection",
+  "veterinary": "veterinary",
+  "videos shop": "videos shop",
+  "video games shop": "video games shop",
+  "viewpoint": "viewpoint",
+  "watches shop": "watches shop",
+  "water park": "water park",
+  "weapons shop": "weapons shop",
+  "warehouse club": "warehouse club",
+  "wine shop": "wine shop",
+  "winery": "winery",
+  "winter sports resort": "winter sports resort",
+  "yoga": "yoga",
+  "zoo": "zoo",
+  "city": "city",
+  "country": "country",
+  "address": "address",
+  "street": "street"
+}
