+++ local_modules/qwant-maps-common/index.js
@@ -1,0 +1,1 @@
+export { getPlaceCategoryName } from './src/place_category_name';
